requirements.txt
flask
paho-mqtt
firebase-admin
gunicorn
